Reset texture parameters on every frame
SimCity 4 is expecting texture parameters to be set as some sort of
global render state, but OpenGL is binding them to whatever texture
happens to be bound at the moment. This is causing textures to be drawn
with the wrong mode sometimes, like retaining walls using GL_CLAMP
instead of the intended GL_REPEAT.

Unfortunately this does mean more GL calls for now, but if we can
upgrade to OpenGL 3.3 and shaders at some point, we can use sampler
objects instead, which do let us manage sampling modes globally.

Fixes #3