Start implementing a basic state tracker
Moves most of the stateful cGDriver commands into a separate class
dedicated to tracking state, and adds some checks in some places to see
if the state has changed before calling into OpenGL.

This doesn't deliver a performance improvement by itself. A lot of this
is also hacked together and a work-in-progress. But this does lay a
foundation to further clean things up and, eventually, implement
multidraw, which is where we should really see things pick up.

See: #8