Crudely implement cross-compiling CMake build system
Wrote this in an afternoon so I could cross-compile from Linux, because I felt like it.

The resulting SCGL.dll is untested at the time of commit, but it will
compile if you have mingw installed. At this time, you must explicitly specify the
compiler if you're not using vscode (or codium) with the CMake extension, which detects
toolchains automagically. Will look into CMake presets if I come back to this.