Add a debug mode window to show game framebuffer
I was trying to debug issues with rendering for dirty rectangles, so I added a
window to show the contents of the framebuffer that backs the dirty rect system.
The color output is wrong because OpenGL's pixels are BGR and I'm just rendering
it to the HWND as RGB, but that doesn't really matter.